move more repositories
