+++ app/src/main/java/com/vultisig/wallet/data/repositories/RepositoriesModule.kt
@@ -1,7 +1,5 @@
 package com.vultisig.wallet.data.repositories
 
-import com.vultisig.wallet.data.db.models.AddressBookOrderEntity
-import com.vultisig.wallet.data.db.models.VaultOrderEntity
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -14,21 +12,9 @@ internal interface RepositoriesModule {
 
     @Binds
     @Singleton
-    fun bindAllowanceRepository(
-        impl: AllowanceRepositoryImpl,
-    ): AllowanceRepository
-
-    @Binds
-    @Singleton
     fun bindChainAccountsRepository(
         impl: AccountsRepositoryImpl,
     ): AccountsRepository
-
-    @Binds
-    @Singleton
-    fun bindBalanceRepositoryImpl(
-        impl: BalanceRepositoryImpl,
-    ): BalanceRepository
 
     @Singleton
     @Binds
@@ -38,97 +24,7 @@ internal interface RepositoriesModule {
 
     @Binds
     @Singleton
-    fun bindGasFeesRepository(
-        impl: GasFeeRepositoryImpl,
-    ): GasFeeRepository
-
-    @Binds
-    @Singleton
-    fun bindDefaultChainsRepository(
-        impl: DefaultChainsRepositoryImpl,
-    ): DefaultChainsRepository
-
-    @Binds
-    @Singleton
-    fun bindBlockChainSpecificRepository(
-        impl: BlockChainSpecificRepositoryImpl,
-    ): BlockChainSpecificRepository
-
-    @Binds
-    @Singleton
-    fun bindExplorerLinkRepository(
-        impl: ExplorerLinkRepositoryImpl,
-    ): ExplorerLinkRepository
-
-    @Binds
-    @Singleton
-    fun bindLastOpenedVaultRepository(
-        impl: LastOpenedVaultRepositoryImpl,
-    ): LastOpenedVaultRepository
-
-    @Binds
-    @Singleton
-    fun bindVaultOrderRepository(
-        impl: VaultOrderRepository
-    ): OrderRepository<VaultOrderEntity>
-
-    @Binds
-    @Singleton
-    fun bindSwapQuoteRepository(
-        impl: SwapQuoteRepositoryImpl
-    ): SwapQuoteRepository
-
-    @Binds
-    @Singleton
     fun bindSwapTransactionRepository(
         impl: SwapTransactionRepositoryImpl
     ): SwapTransactionRepository
-
-    @Binds
-    @Singleton
-    fun bindDepositTransactionRepository(
-        impl: DepositTransactionRepositoryImpl
-    ): DepositTransactionRepository
-
-    @Binds
-    @Singleton
-    fun bindBalanceVisibilityRepository(
-        impl: BalanceVisibilityRepositoryImpl
-    ): BalanceVisibilityRepository
-
-    @Binds
-    @Singleton
-    fun bindAddressBookEntryRepository(
-        impl: AddressBookRepositoryImpl
-    ): AddressBookRepository
-
-    @Binds
-    @Singleton
-    fun bindRequestResultRepository(
-        impl: RequestResultRepositoryImpl
-    ): RequestResultRepository
-
-    @Binds
-    @Singleton
-    fun bindAddressBookOrderRepository(
-        impl: AddressBookOrderRepository
-    ): OrderRepository<AddressBookOrderEntity>
-
-    @Binds
-    @Singleton
-    fun bindSPLTokenRepository(
-        impl: SplTokenRepositoryImpl
-    ): SplTokenRepository
-
-    @Binds
-    @Singleton
-    fun bindBlowfishRepository(
-        impl: BlowfishRepositoryImpl
-    ): BlowfishRepository
-
-    @Binds
-    @Singleton
-    fun bindAddressParserRepository(
-        impl: AddressParserRepositoryImpl
-    ): AddressParserRepository
 }
