+++ app/src/main/java/com/vultisig/wallet/ui/models/home/VaultListViewModel.kt
@@ -3,7 +3,7 @@ package com.vultisig.wallet.ui.models.home
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.vultisig.wallet.data.models.Vault
-import com.vultisig.wallet.data.repositories.VaultOrderRepository
+import com.vultisig.wallet.data.repositories.order.VaultOrderRepository
 import com.vultisig.wallet.data.repositories.VaultRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Job
