+++ app/src/main/java/com/vultisig/wallet/ui/screens/vault_settings/components/ConfirmDeleteViewModel.kt
@@ -8,7 +8,7 @@ import com.vultisig.wallet.data.db.models.VaultOrderEntity
 import com.vultisig.wallet.data.models.calculateAccountsTotalFiatValue
 import com.vultisig.wallet.data.models.calculateAddressesTotalFiatValue
 import com.vultisig.wallet.data.repositories.AccountsRepository
-import com.vultisig.wallet.data.repositories.OrderRepository
+import com.vultisig.wallet.data.repositories.order.OrderRepository
 import com.vultisig.wallet.data.repositories.VaultRepository
 import com.vultisig.wallet.ui.models.mappers.FiatValueToStringMapper
 import com.vultisig.wallet.ui.navigation.Destination
