+++ data/src/main/kotlin/com/vultisig/wallet/data/repositories/BalanceVisibilityRepository.kt
@@ -5,7 +5,7 @@ import com.vultisig.wallet.data.sources.AppDataStore
 import kotlinx.coroutines.flow.first
 import javax.inject.Inject
 
-internal interface BalanceVisibilityRepository {
+interface BalanceVisibilityRepository {
     suspend fun setVisibility(vaultId: String, isVisible: Boolean)
     suspend fun getVisibility(vaultId: String): Boolean
 }
