+++ data/src/main/kotlin/com/vultisig/wallet/data/repositories/DataRepositoriesModule.kt
@@ -1,5 +1,10 @@
 package com.vultisig.wallet.data.repositories
 
+import com.vultisig.wallet.data.db.models.AddressBookOrderEntity
+import com.vultisig.wallet.data.db.models.VaultOrderEntity
+import com.vultisig.wallet.data.repositories.order.AddressBookOrderRepository
+import com.vultisig.wallet.data.repositories.order.OrderRepository
+import com.vultisig.wallet.data.repositories.order.VaultOrderRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -63,4 +68,106 @@ internal interface DataRepositoriesModule {
     fun bindAppLocaleRepository(
         impl: AppLocaleRepositoryImpl,
     ): AppLocaleRepository
+
+    @Binds
+    @Singleton
+    fun bindAddressBookEntryRepository(
+        impl: AddressBookRepositoryImpl
+    ): AddressBookRepository
+
+    @Binds
+    @Singleton
+    fun bindAddressParserRepository(
+        impl: AddressParserRepositoryImpl
+    ): AddressParserRepository
+
+    @Binds
+    @Singleton
+    fun bindAllowanceRepository(
+        impl: AllowanceRepositoryImpl,
+    ): AllowanceRepository
+
+    @Binds
+    @Singleton
+    fun bindBalanceRepositoryImpl(
+        impl: BalanceRepositoryImpl,
+    ): BalanceRepository
+
+    @Binds
+    @Singleton
+    fun bindSPLTokenRepository(
+        impl: SplTokenRepositoryImpl
+    ): SplTokenRepository
+
+    @Binds
+    @Singleton
+    fun bindBalanceVisibilityRepository(
+        impl: BalanceVisibilityRepositoryImpl
+    ): BalanceVisibilityRepository
+
+    @Binds
+    @Singleton
+    fun bindBlockChainSpecificRepository(
+        impl: BlockChainSpecificRepositoryImpl,
+    ): BlockChainSpecificRepository
+
+    @Binds
+    @Singleton
+    fun bindBlowfishRepository(
+        impl: BlowfishRepositoryImpl
+    ): BlowfishRepository
+
+    @Binds
+    @Singleton
+    fun bindDefaultChainsRepository(
+        impl: DefaultChainsRepositoryImpl,
+    ): DefaultChainsRepository
+
+    @Binds
+    @Singleton
+    fun bindDepositTransactionRepository(
+        impl: DepositTransactionRepositoryImpl
+    ): DepositTransactionRepository
+
+    @Binds
+    @Singleton
+    fun bindExplorerLinkRepository(
+        impl: ExplorerLinkRepositoryImpl,
+    ): ExplorerLinkRepository
+
+    @Binds
+    @Singleton
+    fun bindGasFeesRepository(
+        impl: GasFeeRepositoryImpl,
+    ): GasFeeRepository
+
+    @Binds
+    @Singleton
+    fun bindLastOpenedVaultRepository(
+        impl: LastOpenedVaultRepositoryImpl,
+    ): LastOpenedVaultRepository
+
+    @Binds
+    @Singleton
+    fun bindVaultOrderRepository(
+        impl: VaultOrderRepository
+    ): OrderRepository<VaultOrderEntity>
+
+    @Binds
+    @Singleton
+    fun bindAddressBookOrderRepository(
+        impl: AddressBookOrderRepository
+    ): OrderRepository<AddressBookOrderEntity>
+
+    @Binds
+    @Singleton
+    fun bindRequestResultRepository(
+        impl: RequestResultRepositoryImpl
+    ): RequestResultRepository
+
+    @Binds
+    @Singleton
+    fun bindSwapQuoteRepository(
+        impl: SwapQuoteRepositoryImpl
+    ): SwapQuoteRepository
 }
