+++ data/src/main/kotlin/com/vultisig/wallet/data/repositories/LastOpenedVaultRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
-internal interface LastOpenedVaultRepository {
+interface LastOpenedVaultRepository {
 
     val lastOpenedVaultId: Flow<String?>
